Rename fake package to testing package
